requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
streamlit>=1.36.0
scipy>=1.11.0
statsmodels>=0.14.2
pyyaml>=6.0.1
plotly>=5.20.0